requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Upstox SDK
upstox-python-sdk>=2.21.0

# Database
sqlalchemy>=2.0.0

# Scheduling
apscheduler>=3.10.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Authentication
PyJWT>=2.8.0

# WebSocket
websockets>=12.0

# Protobuf (for Upstox market data)
protobuf>=4.25.0
grpcio>=1.60.0

# HTTP
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# Dev / Testing
pytest>=7.4.0
httpx>=0.25.0
